Refactor the Raw struct into a common module and make it identical across Requests and Responses to make it easier to use in shared memory transports.
