+++ src/lib.rs
@@ -64,9 +64,9 @@
 //!#
 //!# let mut stream = MockRead {
 //!#     buffer: vec![
-//!#         0x10, 0xA7, 0xC0, 0x5E, 0x16, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-//!#         0x0, 0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x4, 0x0, 0x1, 0x0, 0x72, 0x6F, 0x6F,
-//!#         0x74
+//!#         0x10, 0xA7, 0xC0, 0x5E, 0x18, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
+//!#         0x0, 0x0, 0x0, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x4, 0x0, 0x1, 0x0, 0x0, 0x0, 0x72,
+//!#         0x6F, 0x6F, 0x74
 //!#     ]
 //!# };
 //!use parsec_interface::operations::{Convert, NativeOperation};
@@ -203,8 +203,8 @@
 //!#
 //!# let mut stream = MockRead {
 //!#     buffer: vec![
-//!#         0x10, 0xA7, 0xC0, 0x5E, 0x14, 0x0, 0x1, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
-//!#         0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x2, 0x0, 0x0, 0x0,
+//!#         0x10, 0xA7, 0xC0, 0x5E, 0x18, 0x0, 0x1, 0x0, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
+//!#         0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x2, 0x0, 0x0, 0x0,
 //!#     ]
 //!# };
 //!use parsec_interface::operations::{Convert, NativeResult};
