+++ src/requests/response/mod.rs
@@ -2,14 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 //! Response definition
 
+use super::common::wire_header_1_0::WireHeader as Raw;
 use super::request::RequestHeader;
 use super::ResponseStatus;
 use super::Result;
 use log::error;
 use std::convert::{TryFrom, TryInto};
 use std::io::{Read, Write};
-
-const RESPONSE_HDR_SIZE: u16 = 20;
 
 mod response_body;
 mod response_header;
@@ -18,7 +17,7 @@ pub use response_body::ResponseBody;
 pub use response_header::ResponseHeader;
 
 #[cfg(feature = "testing")]
-pub use response_header::Raw as RawHeader;
+pub use super::common::wire_header_1_0::WireHeader as RawHeader;
 
 /// Native representation of the response wire format.
 #[derive(PartialEq, Debug)]
@@ -70,7 +69,7 @@ impl Response {
     /// - if encoding any of the fields in the header fails, then
     /// `ResponseStatus::InvalidEncoding` is returned.
     pub fn write_to_stream(self, stream: &mut impl Write) -> Result<()> {
-        let mut raw_header: response_header::Raw = self.header.into();
+        let mut raw_header: Raw = self.header.into();
         raw_header.body_len = u32::try_from(self.body.len())?;
 
         raw_header.write_to_stream(stream)?;
@@ -90,7 +89,7 @@ impl Response {
     /// - if the request body size specified in the header is larger than the limit passed as
     /// a parameter, `BodySizeExceedsLimit` will be returned.
     pub fn read_from_stream(stream: &mut impl Read, body_len_limit: usize) -> Result<Response> {
-        let raw_header = response_header::Raw::read_from_stream(stream)?;
+        let raw_header = Raw::read_from_stream(stream)?;
         let body_len = usize::try_from(raw_header.body_len)?;
         if body_len > body_len_limit {
             error!(
@@ -200,9 +199,9 @@ mod tests {
 
     fn get_response_bytes() -> Vec<u8> {
         vec![
-            0x10, 0xA7, 0xC0, 0x5E, 0x14, 0x00, 0x01, 0x00, 0x00, 0x88, 0x77, 0x66, 0x55, 0x44,
-            0x33, 0x22, 0x11, 0x00, 0x03, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x70, 0x80,
-            0x90,
+            0x10, 0xA7, 0xC0, 0x5E, 0x18, 0x00, 0x01, 0x00, 0x00, 0x88, 0x77, 0x66, 0x55, 0x44,
+            0x33, 0x22, 0x11, 0x00, 0x00, 0x00, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00,
+            0x00, 0x00, 0x70, 0x80, 0x90,
         ]
     }
 }
